Transfert d'argent entre deux comptes : Page JSF et Backing Bean

+++ src/main/java/mg/itu/tpbanquerakotoarison/ejb/GestionnaireCompte.java
@@ -55,7 +55,7 @@ public class GestionnaireCompte {
         return em.merge(compteBancaire);
     }
     
-    public CompteBancaire findById(int id){
+    public CompteBancaire findById(long id){
         return em.find(CompteBancaire.class, id);
     }
 }
